like 50% done with messaging I think

+++ src/main/java/DAO/MessageDAO.java
@@ -1,18 +1,24 @@
 package DAO;
-import Model.Message;
+import Model.*;
 import Util.ConnectionUtil;
 import java.util.List;
 import java.util.ArrayList;
 import java.sql.*;
-
 public class MessageDAO 
 {
     //Process creation of new message
     
-    public void insertMessage(Message message)
+    public Message insertMessage(Message message)
     {
         Connection connection = ConnectionUtil.getConnection();
         try {
+            String checkForAccount = "SELECT * FROM account WHERE account_id = ?";
+            PreparedStatement accountCheckStatement = connection.prepareStatement(checkForAccount);
+            accountCheckStatement.setInt(1, message.posted_by);
+            ResultSet rs = accountCheckStatement.executeQuery();
+
+            if(!rs.next()) return null;
+
             String sql = "INSERT INTO message (posted_by, message_text, time_posted_epoch) VALUES(?, ?, ?)";
 
             PreparedStatement preparedStatement = connection.prepareStatement(sql);
@@ -22,40 +28,44 @@ public class MessageDAO
             preparedStatement.setLong(3, message.time_posted_epoch);
 
             preparedStatement.executeUpdate();
+
+            ResultSet pkeyResultSet = preparedStatement.getGeneratedKeys();
+            if(pkeyResultSet.next()){
+                int generated_message_id = (int) pkeyResultSet.getLong(1);
+                return new Message(generated_message_id, message.posted_by, message.message_text, message.time_posted_epoch);
+            }
         }catch(SQLException e){
             System.out.println(e.getMessage());
         }
+        return null;
     }
 
-    // message_id integer primary key auto_increment,
-    // posted_by integer,
-    // message_text varchar(255),
-    // time_posted_epoch long,
-    // foreign key (posted_by) references Account(account_id)
     //Delete message by ID
-    public void deleteMessageByID(int message_id)
+    public Message deleteMessageByID(int message_id)
     {
         Connection connection = ConnectionUtil.getConnection();
         try {
-            //Write SQL logic here
-            String sql = "DELETE FROM message WHERE message_id = ?";
+            var message = getMessageByID(message_id);
+            if(message == null) return null;
 
+            String sql = "DELETE FROM message WHERE message_id = ?";
             PreparedStatement preparedStatement = connection.prepareStatement(sql);
 
             preparedStatement.setInt(1, message_id);
+            preparedStatement.executeUpdate();
 
-            preparedStatement.executeUpdate();  //TODO: check if correct syntax
+            return message;
         }catch(SQLException e){
             System.out.println(e.getMessage());
         }
+        return null;
     }
 
     //Update message by ID
-    public void updateMessageByID(int message_id, String message_text)
+    public Message updateMessageByID(int message_id, String message_text)
     {
         Connection connection = ConnectionUtil.getConnection();
         try {
-            //Write SQL logic here
             String sql = "UPDATE message SET message_text = ? WHERE message_id = ?";
 
             PreparedStatement preparedStatement = connection.prepareStatement(sql);
@@ -64,9 +74,12 @@ public class MessageDAO
             preparedStatement.setInt(2, message_id);
 
             preparedStatement.executeUpdate();
+
+            return getMessageByID(message_id);  //calculated again to account for the updated text
         }catch(SQLException e){
             System.out.println(e.getMessage());
         }
+        return null;
     }
 
 //Retrieve all messages
@@ -131,7 +144,6 @@ public class MessageDAO
         Connection connection = ConnectionUtil.getConnection();
         List<Message> messages = new ArrayList<>();
         try {
-            //Write SQL logic here
             String sql = "SELECT * FROM message WHERE posted_by = ?";
 
             PreparedStatement preparedStatement = connection.prepareStatement(sql);
